Moved MeasurementTimestamps to timestamp.py.
